Download index wasms in tests instead of at build time
