+++ cycles-ledger/Cargo.toml
@@ -30,10 +30,6 @@ ic-metrics-encoder = "1.1"
 serde_json = "1.0.107"
 leb128 = "0.2.5"
 
-[build-dependencies]
-ureq = "2.9"
-sha2 = "0.10"
-
 [dev-dependencies]
 assert_matches = "1.5.0"
 candid_parser.workspace = true
@@ -47,8 +43,10 @@ icrc1-test-env-state-machine = "0.1.2"
 icrc1-test-suite = "0.1.2"
 lazy_static = "1.4.0"
 proptest = "1.2.0"
+sha2 = "0.10"
 tempfile = "3.10.1"
 tokio = { version = "1.36.0", features = ["rt", "macros"] }
+ureq = "2.9"
 
 [features]
 testing = []
